* 'devel' of https://github.com/arangodb/arangodb: (71 commits)
  backport bug-fix (#9255)
  Bug fix/gravatar ui fix (#9211)
  Bug fix/gharial edge from and tos verification (#9177)
  Bug fix/buffer redux (#9229)
  copy logic of 3.4 scheduler that leaves threads available for FAST lane work (#9235)
  Bug fix/internal issue #589 (#9244)
  upgrade Starter 0.14.5
  Add drop-check for index creation in cluster (#9219)
  wait until task is shut down (#9239)
  attempt to remove expired WAL files on agents (#9231)
  Fix two int size mismatches (#9226)
  fix issues (#9236)
  properly downgrade enterprise features when restoring them into a com… (#9161)
  don't make RocksDB startup fail with 0-byte log files (#9208)
  normalize AQL filter conditions, so that attribute accesses are normally placed on the left hand side of the condition (#9217)
  remove some containers from common.h (#9223)
  Fix shutdown deadlock regarding comm tasks (#9204)
  Revert "Fix Buffer methods (#9070)"
  display reason for error in background listener, plus ignore shutdown (#9206)
  don't include the Logger in header files if it's not necessary (#9216)
  ...

+++ tests/Aql/HashedCollectExecutorTest.cpp
@@ -192,7 +192,7 @@ TEST_F(HashedCollectExecutorTestRowsNoCount, the_producer_doesnt_wait_1) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   auto block = result.stealBlock();
 
   // check for types
@@ -237,7 +237,7 @@ TEST_F(HashedCollectExecutorTestRowsNoCount, the_producer_doesnt_wait_2) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   auto block = result.stealBlock();
 
   // check for types
@@ -287,7 +287,7 @@ TEST_F(HashedCollectExecutorTestRowsNoCount, the_producer_doesnt_wait_3) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   auto block = result.stealBlock();
 
   // check for types
@@ -332,7 +332,7 @@ TEST_F(HashedCollectExecutorTestRowsNoCount, the_producer_doesnt_wait_4) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   auto block = result.stealBlock();
 
   // check for types
@@ -380,7 +380,7 @@ TEST_F(HashedCollectExecutorTestRowsNoCount, the_producer_waits) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   auto block = result.stealBlock();
 
   // check for types
@@ -462,7 +462,7 @@ TEST(HashedCollectExecutorTestRowsCount, the_producer_doesnt_wait) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
+  std::vector<int64_t> myNumbers;
   std::vector<double> myCountNumbers;
   auto newBlock = result.stealBlock();
 
@@ -564,8 +564,8 @@ TEST(HashedCollectExecutorTestRowsCountNumbers, the_producer_doesnt_wait) {
   ASSERT_TRUE(state == ExecutionState::DONE);
   ASSERT_TRUE(!result.produced());
 
-  std::vector<int> myNumbers;
-  std::vector<int> myCountNumbers;
+  std::vector<int64_t> myNumbers;
+  std::vector<int64_t> myCountNumbers;
   auto newBlock = result.stealBlock();
 
   // check for types
@@ -678,7 +678,7 @@ TEST(HashedCollectExecutorTestRowsCountStrings, the_producer_doesnt_wait) {
   ASSERT_TRUE(!result.produced());
 
   std::vector<std::string> myStrings;
-  std::vector<int> myCountNumbers;
+  std::vector<int64_t> myCountNumbers;
   auto newBlock = result.stealBlock();
 
   // check for types
